requirements-dev.txt
Django==1.11.1
django-material
